+++ src/cubyz/rendering/rotation/TorchRotation.java
@@ -2,9 +2,10 @@ package cubyz.rendering.rotation;
 
 import org.joml.Matrix3f;
 import org.joml.RayAabIntersection;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 import org.joml.Vector3i;
-import org.joml.Vector4f;
+import org.joml.Vector4d;
 
 import cubyz.api.Resource;
 import cubyz.client.Meshes;
@@ -38,7 +39,7 @@ public class TorchRotation implements RotationMode {
 	}
 
 	@Override
-	public boolean generateData(ServerWorld world, int x, int y, int z, Vector3f relativePlayerPosition, Vector3f playerDirection, Vector3i relativeDirection, ByteWrapper currentData, boolean blockPlacing) {
+	public boolean generateData(ServerWorld world, int x, int y, int z, Vector3d relativePlayerPosition, Vector3f playerDirection, Vector3i relativeDirection, ByteWrapper currentData, boolean blockPlacing) {
 		byte data = (byte)0;
 		if(relativeDirection.x == 1) data = (byte)0b1;
 		if(relativeDirection.x == -1) data = (byte)0b10;
@@ -109,12 +110,12 @@ public class TorchRotation implements RotationMode {
 	}
 
 	@Override
-	public boolean checkEntity(Vector3f pos, float width, float height, int x, int y, int z, byte blockData) {
+	public boolean checkEntity(Vector3d pos, double width, double height, int x, int y, int z, byte blockData) {
 		return false;
 	}
 
 	@Override
-	public boolean checkEntityAndDoCollision(Entity arg0, Vector4f arg1, int x, int y, int z, byte arg2) {
+	public boolean checkEntityAndDoCollision(Entity arg0, Vector4d arg1, int x, int y, int z, byte arg2) {
 		return true;
 	}
